requirements.txt
# Trading y APIs
ccxt==4.2.25
python-binance==1.0.19
websocket-client==1.7.0

# Análisis Técnico y Datos
pandas==2.2.0
numpy==1.26.3
# ta-lib==0.4.28  # Comentado temporalmente - problemas de compilación
yfinance==0.2.18

# API y Web Framework
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
aiohttp==3.9.1

# Base de Datos y Cache
redis==5.0.1
aioredis==2.0.1

# Visualización y Dashboard
plotly==5.18.0
dash==2.16.1
dash-bootstrap-components==1.5.0

# Utilidades
python-dotenv==1.0.0
pydantic==2.5.3
loguru==0.7.2
asyncio==3.4.3
requests==2.31.0

# Machine Learning (opcional)
scikit-learn==1.4.0
